requirements.txt
# requirements.txt
streamlit==1.32.0
google-generativeai==0.5.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.1
faiss-cpu==1.8.0
numpy==1.26.4